Hierarchysort compiles but crashes

+++ buoyancysort/fractal-max-heap.h
@@ -11,20 +11,20 @@ namespace FractalMaxHeap
 	template <typename Type>
 	Type parent(Type child, Type center, double tuple_size)
 	{
-		return floor((child - center) / tuple_size) + center;
+		return (long) floor((child - center) / tuple_size) + center;
 	}
 
 	/// -1/0 - based
 	template <typename Type>
-	Type left_child(Type parent, Type center, long tuple_size)
+	Type left_child(Type parent, Type center, double tuple_size)
 	{
-		return ceil((parent - center + 1) * tuple_size - (tuple_size - 1)) + center;
+		return (long) ceil((parent - center + 1) * tuple_size - (tuple_size - 1)) + center;
 	}
 
 	/// -1/0 - based
 	template <typename Type>
-	Type right_child(Type parent, Type center, long tuple_size)
+	Type right_child(Type parent, Type center, double tuple_size)
 	{
-		return ceil((parent - center + 1) * tuple_size) + center;
+		return (long) ceil((parent - center + 1) * tuple_size) + center;
 	}
 }
